add task crud

+++ src/main/java/hexlet/code/service/TaskStatusService.java
@@ -9,6 +9,8 @@ import hexlet.code.dto.TaskStatusUpdateDTO;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.TaskStatusMapper;
 import hexlet.code.repository.TaskStatusRepository;
+import hexlet.code.repository.TaskRepository;
+import hexlet.code.exception.BadRequestException;
 
 import java.util.List;
 
@@ -20,6 +22,9 @@ public class TaskStatusService {
 
     @Autowired
     private TaskStatusMapper mapper;
+
+    @Autowired
+    private TaskRepository taskRepository;
 
     public List<TaskStatusDTO> getAll() {
         var taskStatuses = repository.findAll();
@@ -51,6 +56,13 @@ public class TaskStatusService {
     }
 
     public void delete(Long id) {
+        var taskStatus = repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Task status with id " + id + " not found"));
+
+        if (!taskRepository.findByTaskStatusSlug(taskStatus.getSlug()).isEmpty()) {
+            throw new BadRequestException("Cannot delete the task status as it is used in tasks");
+        }
+
         repository.deleteById(id);
     }
 }
